replace parseInt() to parseFloat in getTopCorrection()

+++ extension/src/content/bundle/ipa.js
@@ -134,7 +134,7 @@
 	});
 
 	function getTopCorrection() {
-		return parseInt(sizepx) * 1.8;
+		return parseFloat(sizepx) * 1.8;
 	}
 
 	function backgroundColor() {
